* 'master' of https://github.com/TrinityCore/TrinityCore:
  Core/Entities: Allow 2 side parties to work.
  Core/Entities: GameObject's m_unique_users will now store full guids to get rid of silly casts
  Core/Grids: Added GameObjectWorker
  Core/Commands: Removed setting orientation from cast commands
  Core/Entities: Allow creatures to attack players while in sanctuary.
  Core/AI: Replace many Unit::IsHostileTo with Unit::IsValidAttackTarget or Creature::canCreatureAttack.
  Core/Entities: Remove obsolete function Unit::canAttack.
  Core/Entities: Replace Unit::canAttack usage with more accurate Unit::IsValidAttackTarget
  Core/Misc: Whitespace and a minor typo fix (thanks Kapoeira)
  Core/Spells: Use Unit::IsValidAttack/AssistTarget functions to validate spell explicit targets.
  Core/Entities: Add Unit::IsValidAssistTarget and Unit::IsValidAttackTarget functions for proper serverside checks of related entity flags.
  Tab2spaces previous commit :S
  Core/Instances: Resurrect a dead player (with corpse) before enter to instance. Original patch by Den
  Core/Instances: Fix client freeze when trying to enter an instance which have players inside and you have save but not in same group that people inside and your homebind is in the same continent that the instance

+++ src/server/game/Entities/GameObject/GameObject.h
@@ -800,7 +800,7 @@ class GameObject : public WorldObject, public GridObject<GameObject>
         std::list<uint32> m_SkillupList;
 
         Player* m_ritualOwner;                              // used for GAMEOBJECT_TYPE_SUMMONING_RITUAL where GO is not summoned (no owner)
-        std::set<uint32> m_unique_users;
+        std::set<uint64> m_unique_users;
         uint32 m_usetimes;
 
         typedef std::map<uint32, uint64> ChairSlotAndUser;
